+++ test/diagnostics_test.rb
@@ -52,14 +52,14 @@ class DiagnosticsTest < Minitest::Test
 
   def test_performance_metrics_invalid
     error = assert_raises(ArgumentError) do
-      Prophet::Diagnostics.performance_metrics(Rover::DataFrame.new, metrics: ["invalid"])
+      Prophet::Diagnostics.performance_metrics(Polars::DataFrame.new, metrics: ["invalid"])
     end
     assert_match "Valid values for metrics are: ", error.message
   end
 
   def test_performance_metrics_non_unique
     error = assert_raises(ArgumentError) do
-      Prophet::Diagnostics.performance_metrics(Rover::DataFrame.new, metrics: ["mse", "mse"])
+      Prophet::Diagnostics.performance_metrics(Polars::DataFrame.new, metrics: ["mse", "mse"])
     end
     assert_equal "Input metrics must be a list of unique values", error.message
   end
@@ -88,7 +88,7 @@ class DiagnosticsTest < Minitest::Test
     end
 
     # Find the best parameters
-    tuning_results = Rover::DataFrame.new(all_params)
+    tuning_results = Polars::DataFrame.new(all_params)
     tuning_results["rmse"] = rmses
     assert_equal 16, tuning_results.size
   end