+++ test/support/custom_cutoffs.py
@@ -2,7 +2,7 @@ import pandas as pd
 from prophet import Prophet
 from prophet.diagnostics import cross_validation, performance_metrics
 
-# float_precision='high' required for pd.read_csv to match precision of Rover.read_csv
+# float_precision='high' required for pd.read_csv to match precision of Polars.read_csv
 df = pd.read_csv('examples/example_wp_log_peyton_manning.csv', float_precision='high')
 
 m = Prophet()